+++ Runtime/Components/InstanceGuid.cs
@@ -1,0 +1,24 @@
+﻿using GameCreator.Runtime.Common;
+using UnityEngine;
+
+namespace GameCreator.Runtime.SaveablePrefabs
+{
+    [DisallowMultipleComponent]
+    public class InstanceGuid : MonoBehaviour
+    {
+        [SerializeField]
+        UniqueID _guid = new();
+
+        public int GuidHash => _guid.Get.Hash;
+
+        void Awake()
+        {
+            SaveablePrefabInstanceManager.RegisterInstanceGuid(this);
+        }
+
+        void OnDestroy()
+        {
+            SaveablePrefabInstanceManager.UnregisterInstanceGuid(this);
+        }
+    }
+}